AppData/Roaming/AbacusAI/Agent Workspaces/requirements.txt
selenium
webdriver-manager
flask
flask-cors
PyPDF2
requests
opencv-python
pillow
numpy
